requirements.txt
google-cloud-alloydb-connector[asyncpg]==1.5.0
llama-index-core==0.12.3
pgvector==0.3.6
SQLAlchemy[asyncio]==2.0.36
